+++ Assets/Features/Player/Scripts/BasePlayerLocomotion.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 68d1c1018df3a3b4e83e176d5c9d408c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
