+++ Assets/Features/Player/Scripts/CameraLocomotionFeature.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e24fdde439633434e947257f72a7ef8e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
