+++ Assets/Features/Player/Scripts/CoyoteTimeJumpLocomotionFeature.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 657181d04e5acd54cb984076941dd86e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
