+++ Assets/Features/Player/Scripts/MoveLocomotionFeature.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2b71dc697c142b4459cddd00779c0316
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
