+++ Assets/Features/Player/Scripts/PlayerLocomotionFeature.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aae99948114ae304bbb524ec0cb8fdff
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
